requirements.txt
discord.py
git+git://github.com/kennethreitz/requests.git